+++ packages/http_client_contracts/lib/src/http_cancellation_token.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
-import 'package:http_client_core/src/http_exception.dart';
-import 'package:http_client_core/src/http_request.dart';
+import 'package:http_client_contracts/src/http_exception.dart';
+import 'package:http_client_contracts/src/http_request.dart';
 
 class HttpCancellationToken {
   bool _isCancelled = false;