+++ packages/http_client_contracts/lib/src/http_stream_response.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:typed_data';
 
-import 'package:http_client_core/src/http_request.dart';
+import 'package:http_client_contracts/src/http_request.dart';
 
 class HttpStreamResponse {
   final HttpRequest request;