+++ packages/http_client_http/lib/http_client_http.dart
@@ -1,3 +1,3 @@
-export 'package:http_client_core/http_client_core.dart';
+export 'package:http_client_contracts/http_client_contracts.dart';
 
 export 'src/package_http_client.dart';